move test to phase0
