+++ tests/core/pyspec/eth2spec/test/phase0/sanity/test_slots.py
@@ -6,6 +6,8 @@ from eth2spec.test.context import (
     spec_state_test,
     with_all_phases,
 )
+from eth2spec.test.helpers.state import next_epoch
+from eth2spec.test.helpers.state import transition_to
 
 
 @with_all_phases
@@ -87,3 +89,34 @@ def test_historical_accumulator(spec, state):
         assert len(state.historical_summaries) == len(pre_historical_summaries) + 1
     else:
         assert len(state.historical_roots) == len(pre_historical_roots) + 1
+
+
+def run_epoch_processing(spec, state):
+    slot = state.slot + spec.SLOTS_PER_EPOCH - (state.slot % spec.SLOTS_PER_EPOCH) - 1
+    transition_to(spec, state, slot)
+    yield 'pre', state
+    yield 'slots', 1
+    spec.process_slots(state, state.slot + 1)
+    yield 'post', state
+
+
+@with_all_phases
+@spec_state_test
+def test_balance_change_affects_proposer(spec, state):
+    # Get 1 epoch in the future and get the proposer index selected
+    future_state = state.copy()
+    next_epoch(spec, future_state)
+    proposer_next_epoch = spec.get_beacon_proposer_index(future_state)
+
+    # Set the effective balance of the proposer selected to 0
+    state.validators[proposer_next_epoch].effective_balance = 0
+
+    # Process the epoch to go forward in time and ensure the balance
+    # change is taken into account
+    run_epoch_processing(spec, state)
+
+    # Retrieve the index of the proposer selected
+    proposer_next_epoch_after_change = spec.get_beacon_proposer_index(state)
+
+    # Assert that the proposer selected did change
+    assert proposer_next_epoch != proposer_next_epoch_after_change
